Project One
This project had one goal in mind, Create a text.txt file read that file and determine the number of upper case lower case and numbers are in the text file itself. It was pretty easy and even had a built in loop so you can open save and rerun the program without a total restart of the program. Pretty nifty and fun project to get started in week one of Programming Fundamentals two.